requirements.txt
bs4
requests
json
python-telegram-bot
